refactor(api-event-handler): rename api-infra-* → api-event-handler-*
Rename the API composition packages to parallel the event-handler-* transport
layer they build on:
- @webiny/api-infra-aws        → @webiny/api-event-handler-aws
- @webiny/api-infra-aws-ddb    → @webiny/api-event-handler-aws-ddb
- @webiny/api-infra-aws-ddb-os → @webiny/api-event-handler-aws-ddb-os

Pure rename: dir moves + package names + all imports + template shims +
comments; regenerated tsconfigs/references/lockfile. All three build.
(api-event-handler-core extraction + template-dep prune follow in separate commits.)
