+++ packages/api-event-handler-aws-ddb-os/src/index.ts
