+++ packages/api-event-handler-aws-ddb-os/tsconfig.build.json
@@ -5,8 +5,8 @@
     { "path": "../api-aco-ddb/tsconfig.build.json" },
     { "path": "../api-audit-logs-ddb/tsconfig.build.json" },
     { "path": "../api-core-ddb/tsconfig.build.json" },
+    { "path": "../api-event-handler-aws/tsconfig.build.json" },
     { "path": "../api-headless-cms-ddb-es/tsconfig.build.json" },
-    { "path": "../api-infra-aws/tsconfig.build.json" },
     { "path": "../api-opensearch/tsconfig.build.json" },
     { "path": "../api-websockets-ddb/tsconfig.build.json" },
     { "path": "../db/tsconfig.build.json" }
@@ -24,10 +24,10 @@
       "@webiny/api-audit-logs-ddb": ["../api-audit-logs-ddb/src"],
       "@webiny/api-core-ddb/*": ["../api-core-ddb/src/*"],
       "@webiny/api-core-ddb": ["../api-core-ddb/src"],
+      "@webiny/api-event-handler-aws/*": ["../api-event-handler-aws/src/*"],
+      "@webiny/api-event-handler-aws": ["../api-event-handler-aws/src"],
       "@webiny/api-headless-cms-ddb-es/*": ["../api-headless-cms-ddb-es/src/*"],
       "@webiny/api-headless-cms-ddb-es": ["../api-headless-cms-ddb-es/src"],
-      "@webiny/api-infra-aws/*": ["../api-infra-aws/src/*"],
-      "@webiny/api-infra-aws": ["../api-infra-aws/src"],
       "@webiny/api-opensearch/*": ["../api-opensearch/src/*"],
       "@webiny/api-opensearch": ["../api-opensearch/src"],
       "@webiny/api-websockets-ddb/*": ["../api-websockets-ddb/src/*"],
