+++ packages/api-event-handler-aws-ddb-os/webiny.config.js
