+++ packages/api-event-handler-aws-ddb/jest-dynalite-config.cjs
