+++ packages/api-event-handler-aws-ddb/src/index.ts
