+++ packages/api-event-handler-aws-ddb/tsconfig.build.json
@@ -5,8 +5,8 @@
     { "path": "../api-aco-ddb/tsconfig.build.json" },
     { "path": "../api-audit-logs-ddb/tsconfig.build.json" },
     { "path": "../api-core-ddb/tsconfig.build.json" },
+    { "path": "../api-event-handler-aws/tsconfig.build.json" },
     { "path": "../api-headless-cms-ddb/tsconfig.build.json" },
-    { "path": "../api-infra-aws/tsconfig.build.json" },
     { "path": "../api-websockets-ddb/tsconfig.build.json" }
   ],
   "compilerOptions": {
@@ -22,10 +22,10 @@
       "@webiny/api-audit-logs-ddb": ["../api-audit-logs-ddb/src"],
       "@webiny/api-core-ddb/*": ["../api-core-ddb/src/*"],
       "@webiny/api-core-ddb": ["../api-core-ddb/src"],
+      "@webiny/api-event-handler-aws/*": ["../api-event-handler-aws/src/*"],
+      "@webiny/api-event-handler-aws": ["../api-event-handler-aws/src"],
       "@webiny/api-headless-cms-ddb/*": ["../api-headless-cms-ddb/src/*"],
       "@webiny/api-headless-cms-ddb": ["../api-headless-cms-ddb/src"],
-      "@webiny/api-infra-aws/*": ["../api-infra-aws/src/*"],
-      "@webiny/api-infra-aws": ["../api-infra-aws/src"],
       "@webiny/api-websockets-ddb/*": ["../api-websockets-ddb/src/*"],
       "@webiny/api-websockets-ddb": ["../api-websockets-ddb/src"]
     }
