+++ packages/api-event-handler-aws-ddb/tsconfig.json
@@ -5,11 +5,9 @@
     { "path": "../api-aco-ddb" },
     { "path": "../api-audit-logs-ddb" },
     { "path": "../api-core-ddb" },
-    { "path": "../api-headless-cms-ddb-es" },
-    { "path": "../api-infra-aws" },
-    { "path": "../api-opensearch" },
-    { "path": "../api-websockets-ddb" },
-    { "path": "../db" }
+    { "path": "../api-event-handler-aws" },
+    { "path": "../api-headless-cms-ddb" },
+    { "path": "../api-websockets-ddb" }
   ],
   "compilerOptions": {
     "rootDirs": ["./src", "./__tests__"],
@@ -24,16 +22,12 @@
       "@webiny/api-audit-logs-ddb": ["../api-audit-logs-ddb/src"],
       "@webiny/api-core-ddb/*": ["../api-core-ddb/src/*"],
       "@webiny/api-core-ddb": ["../api-core-ddb/src"],
-      "@webiny/api-headless-cms-ddb-es/*": ["../api-headless-cms-ddb-es/src/*"],
-      "@webiny/api-headless-cms-ddb-es": ["../api-headless-cms-ddb-es/src"],
-      "@webiny/api-infra-aws/*": ["../api-infra-aws/src/*"],
-      "@webiny/api-infra-aws": ["../api-infra-aws/src"],
-      "@webiny/api-opensearch/*": ["../api-opensearch/src/*"],
-      "@webiny/api-opensearch": ["../api-opensearch/src"],
+      "@webiny/api-event-handler-aws/*": ["../api-event-handler-aws/src/*"],
+      "@webiny/api-event-handler-aws": ["../api-event-handler-aws/src"],
+      "@webiny/api-headless-cms-ddb/*": ["../api-headless-cms-ddb/src/*"],
+      "@webiny/api-headless-cms-ddb": ["../api-headless-cms-ddb/src"],
       "@webiny/api-websockets-ddb/*": ["../api-websockets-ddb/src/*"],
-      "@webiny/api-websockets-ddb": ["../api-websockets-ddb/src"],
-      "@webiny/db/*": ["../db/src/*"],
-      "@webiny/db": ["../db/src"]
+      "@webiny/api-websockets-ddb": ["../api-websockets-ddb/src"]
     }
   }
 }
