+++ packages/api-event-handler-aws-ddb/vitest.config.ts
