+++ packages/api-event-handler-aws-ddb/webiny.config.js
