+++ packages/api-event-handler-aws/__tests__/S3TenantLoaderDecorator.test.ts
