+++ packages/api-event-handler-aws/src/handlers/S3TenantLoaderDecorator.ts
