+++ packages/api-event-handler-aws/src/handlers/index.ts
