+++ packages/api-event-handler-aws/src/index.ts
