+++ packages/api-event-handler-aws/tsconfig.build.json
