+++ packages/api-event-handler-aws/tsconfig.json
