+++ packages/project-aws/_templates/extensions/OpenSearch/api/graphql/src/index.ts
@@ -1,12 +1,12 @@
 /**
  * DI-native API GraphQL handler — OpenSearch (DDB+ES) variant.
  *
- * The composition root lives in @webiny/api-infra-aws-ddb-os (`createWebinyApiHandler`) so the
+ * The composition root lives in @webiny/api-event-handler-aws-ddb-os (`createWebinyApiHandler`) so the
  * wiring is a real, testable package rather than template code. The OpenSearch extension's
  * ReplaceApiLambdaFnHandlers copies this file over the api workspace when OpenSearch is enabled.
  * This file only supplies project-specific extensions.
  */
-import { createWebinyApiHandler } from "@webiny/api-infra-aws-ddb-os";
+import { createWebinyApiHandler } from "@webiny/api-event-handler-aws-ddb-os";
 import { extensions } from "./extensions";
 
 export const handler = createWebinyApiHandler({ extensions });
